+++ src/tcp/method_traits.rs
@@ -196,7 +196,7 @@ pub trait TcpMethods: HeaderInformation + TcpUdpChecksum + BufferAccess {
 
     #[inline]
     fn tcp_calculate_checksum(&self) -> u16 {
-        let checksum = self.pseudoheader_checksum(self.data_buffer_starting_at_header(LAYER).len());
+        let checksum = self.pseudoheader_checksum();
 
         internet_checksum::<4>(checksum, self.data_buffer_starting_at_header(LAYER))
     }