+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - TCP docs ([62cb561](https://github.com/arctic-alpaca/mutnet/commit/62cb5614b819d304bd38c3239d34c3cfff07f500) & [9601e0b](https://github.com/arctic-alpaca/mutnet/commit/9601e0bd56c1ee3b023453573c4fa830362b1ea3)) 
+- TCP/UDP checksum now uses information from the IPv4/IPv6 header in the pseudo header
 
 # 0.1.0 (19. September, 2023)
 ### Added